修复无法使用拆分视图的问题,优化快速启动显示效果
