+++ RX_Explorer/Class/CommonAccessCollection.cs
@@ -68,6 +68,8 @@ namespace RX_Explorer.Class
                                 HotWebList.Add(Item.Value);
                             }
                         }
+                        QuickStartList.Add(new QuickStartItem());
+                        HotWebList.Add(new QuickStartItem());
                     }
                 }
                 finally
