+++ RX_Explorer/Class/QuickStartItem.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using Windows.UI.Xaml.Media.Imaging;
 
 namespace RX_Explorer.Class
@@ -79,6 +80,12 @@ namespace RX_Explorer.Class
 
             this.DisplayName = DisplayName;
             this.RelativePath = RelativePath;
+        }
+
+        public QuickStartItem()
+        {
+            Type = QuickStartType.AddButton;
+            Image = new BitmapImage(new Uri("ms-appx:///Assets/AddImage.png"));
         }
     }
 }
