+++ RX_Explorer/View/FileControl.xaml.cs
@@ -129,8 +129,6 @@ namespace RX_Explorer
 
             BladePointerPressedEventHandler = new PointerEventHandler(Blade_PointerPressed);
 
-            LibraryExpander.IsExpanded = SettingControl.LibraryExpanderIsExpand;
-            DeviceExpander.IsExpanded = SettingControl.DeviceExpanderIsExpand;
 
             Loaded += FileControl_Loaded;
         }
@@ -451,7 +449,7 @@ namespace RX_Explorer
         private string[] GetPathSplit(string Path)
         {
             
-            if(Path!= Globalization.GetString("MainPage_PageDictionary_ThisPC_Label")&&!Path.StartsWith("@\\"))
+            if(Path != Globalization.GetString("MainPage_PageDictionary_ThisPC_Label")&&!Path.StartsWith("@\\"))
             {
                 Path = Globalization.GetString("MainPage_PageDictionary_ThisPC_Label") + @"\" + Path;
             }
@@ -484,10 +482,7 @@ namespace RX_Explorer
                          
                         Frame.Navigated += Frame_Navigated;
                         FullTrustProcessController.CurrentBusyStatus += FullTrustProcessController_CurrentBusyStatus;
-                        CommonAccessCollection.DriveAdded += CommonAccessCollection_DriveAdded;
-                        CommonAccessCollection.DriveRemoved += CommonAccessCollection_DriveRemoved;
 
-                           
 
                         ViewModeControl = new ViewModeController();
 
@@ -503,6 +498,8 @@ namespace RX_Explorer
                         
                         await Initialize(Parameters.Item2).ConfigureAwait(false);
                     }
+                    CommonAccessCollection.DriveAdded += CommonAccessCollection_DriveAdded;
+                    CommonAccessCollection.DriveRemoved += CommonAccessCollection_DriveRemoved;
                 }
                 catch (Exception ex)
                 {
@@ -568,10 +565,11 @@ namespace RX_Explorer
         {
 
             if (InitFolderPathArray.Length > 0)
-            { 
-                 
+            {
+                
                 foreach (string TargetPath in InitFolderPathArray.Where((FolderPath) => !string.IsNullOrWhiteSpace(FolderPath)))
                 {
+
                     await CreateNewBladeAsync(TargetPath);
                 }
                 DriveDataBase[] Drives = CommonAccessCollection.DriveList.Where((Drive) => !string.IsNullOrWhiteSpace(Drive.Path)).ToArray();
@@ -597,7 +595,7 @@ namespace RX_Explorer
                     }
                 }
 
-                
+
 
                 foreach (DriveDataBase DriveData in Drives.Where((Dr) => Dr.DriveType == DriveType.Network))
                 {
@@ -623,11 +621,11 @@ namespace RX_Explorer
 
                 //if(CurrentPresenter.CurrentFolder.Path == Globalization.GetString("MainPage_PageDictionary_ThisPC_Label"))
                 //{
-                 
+
                 //}
                 //else { 
                 //    string TargetRootPath = Path.GetPathRoot(CurrentPresenter.CurrentFolder.Path);
-                
+
                 //    if (FolderTree.RootNodes.FirstOrDefault((Node) => (Node.Content as TreeViewNodeContent).Path == TargetRootPath) is TreeViewNode TargetRootNode)
                 //    {
                 //        FolderTree.SelectNodeAndScrollToVertical(TargetRootNode);
@@ -1535,7 +1533,7 @@ namespace RX_Explorer
                         }
                         else if (!string.IsNullOrWhiteSpace(DirectoryPath))
                         {
-                            HideThisPC();
+                            
                             await CurrentPresenter.DisplayItemsInFolder(DirectoryPath);
 
                             if (CurrentPresenter.FileCollection.OfType<FileSystemStorageFolder>().FirstOrDefault((Item) => Item.Path.Equals(CurrentFolderPath, StringComparison.OrdinalIgnoreCase)) is FileSystemStorageItemBase Folder)
@@ -2665,602 +2663,6 @@ namespace RX_Explorer
                 await SQLite.Current.DeletePathHistoryAsync(Item.Path);
             }
         }
-        private void LibraryExpander_Collapsed(object sender, EventArgs e)
-        {
-            SettingControl.LibraryExpanderIsExpand = false;
-            LibraryGrid.SelectedIndex = -1;
-        }
-
-        private void DeviceExpander_Collapsed(object sender, EventArgs e)
-        {
-            SettingControl.DeviceExpanderIsExpand = false;
-            DeviceGrid.SelectedIndex = -1;
-        }
-
-        private async void EjectButton_Click(object sender, RoutedEventArgs e)
-        {
-            if (DeviceGrid.SelectedItem is DriveDataBase Item)
-            {
-                if (string.IsNullOrEmpty(Item.Path))
-                {
-                    QueueContentDialog Dialog = new QueueContentDialog
-                    {
-                        Title = Globalization.GetString("Common_Dialog_ErrorTitle"),
-                        Content = Globalization.GetString("QueueContentDialog_UnableToEject_Content"),
-                        CloseButtonText = Globalization.GetString("Common_Dialog_CloseButton")
-                    };
-
-                    await Dialog.ShowAsync().ConfigureAwait(false);
-                }
-                else
-                {
-                    foreach (TabViewItem Tab in TabViewContainer.ThisPage.TabCollection.Where((Tab) => (Tab.Content as Frame).CurrentSourcePageType != typeof(ThisPC) && Tab.Tag is FileControl Control && Path.GetPathRoot(Control.CurrentPresenter.CurrentFolder?.Path).Equals(Item.Path, StringComparison.OrdinalIgnoreCase)).ToArray())
-                    {
-                        await TabViewContainer.ThisPage.CleanUpAndRemoveTabItem(Tab);
-                    }
-
-                    using (FullTrustProcessController.ExclusiveUsage Exclusive = await FullTrustProcessController.GetAvailableController())
-                    {
-                        if (await Exclusive.Controller.EjectPortableDevice(Item.Path))
-                        {
-                            ShowEjectNotification();
-                        }
-                        else
-                        {
-                            QueueContentDialog Dialog = new QueueContentDialog
-                            {
-                                Title = Globalization.GetString("Common_Dialog_ErrorTitle"),
-                                Content = Globalization.GetString("QueueContentDialog_UnableToEject_Content"),
-                                CloseButtonText = Globalization.GetString("Common_Dialog_CloseButton")
-                            };
-
-                            await Dialog.ShowAsync().ConfigureAwait(false);
-                        }
-                    }
-                }
-            }
-        }
-
-        private void ShowEjectNotification()
-        {
-            try
-            {
-                ToastNotificationManager.History.Remove("MergeVideoNotification");
-
-                ToastContentBuilder Builder = new ToastContentBuilder()
-                                              .SetToastScenario(ToastScenario.Default)
-                                              .AddToastActivationInfo("Transcode", ToastActivationType.Foreground)
-                                              .AddText(Globalization.GetString("Eject_Toast_Text_1"))
-                                              .AddText(Globalization.GetString("Eject_Toast_Text_2"))
-                                              .AddText(Globalization.GetString("Eject_Toast_Text_3"));
-
-                ToastNotificationManager.CreateToastNotifier().Show(new ToastNotification(Builder.GetToastContent().GetXml()));
-            }
-            catch (Exception ex)
-            {
-                LogTracer.Log(ex, "Toast notification could not be sent");
-            }
-        }
-        private async void LibraryGrid_DragItemsCompleted(ListViewBase sender, DragItemsCompletedEventArgs args)
-        {
-            await SQLite.Current.ClearTableAsync("Library");
-
-            foreach (LibraryFolder Item in CommonAccessCollection.LibraryFolderList)
-            {
-                await SQLite.Current.SetLibraryPathAsync(Item.Folder.Path, Item.Type);
-            }
-        }
-
-        private async void AddLibraryButton_Click(object sender, RoutedEventArgs e)
-        {
-            FolderPicker Picker = new FolderPicker
-            {
-                ViewMode = PickerViewMode.List,
-                SuggestedStartLocation = PickerLocationId.ComputerFolder
-            };
-            Picker.FileTypeFilter.Add("*");
-
-            if (await Picker.PickSingleFolderAsync() is StorageFolder Folder)
-            {
-                if (CommonAccessCollection.LibraryFolderList.Any((Library) => Library.Folder.Path.Equals(Folder.Path, StringComparison.OrdinalIgnoreCase)))
-                {
-                    QueueContentDialog dialog = new QueueContentDialog
-                    {
-                        Title = Globalization.GetString("Common_Dialog_TipTitle"),
-                        Content = Globalization.GetString("QueueDialog_RepeatAddToHomePage_Content"),
-                        CloseButtonText = Globalization.GetString("Common_Dialog_CloseButton")
-                    };
-                    await dialog.ShowAsync();
-                }
-                else
-                {
-                    CommonAccessCollection.LibraryFolderList.Add(await LibraryFolder.CreateAsync(Folder, LibraryType.UserCustom));
-                    await SQLite.Current.SetLibraryPathAsync(Folder.Path, LibraryType.UserCustom).ConfigureAwait(false);
-                    await JumpListController.Current.AddItemAsync(JumpListGroup.Library, Folder.Path).ConfigureAwait(false);
-                }
-            }
-        }
-
-        private async void UnlockBitlocker_Click(object sender, RoutedEventArgs e)
-        {
-            if (DeviceGrid.SelectedItem is DriveDataBase Drive)
-            {
-                await OpenTargetDriveAsync(Drive);
-            }
-        }
-
-        private void GridView_PreviewKeyDown(object sender, Windows.UI.Xaml.Input.KeyRoutedEventArgs e)
-        {
-            if (e.Key == VirtualKey.Space)
-            {
-                e.Handled = true;
-            }
-        }
-
-        private async void LibraryExpander_Expanded(object sender, EventArgs e)
-        {
-            SettingControl.LibraryExpanderIsExpand = true;
-            await CommonAccessCollection.LoadLibraryFoldersAsync();
-        }
-
-        private async void DeviceExpander_Expanded(object sender, EventArgs e)
-        {
-            SettingControl.DeviceExpanderIsExpand = true;
-            await CommonAccessCollection.LoadDriveAsync();
-        }
-
-        private void DeviceGrid_ContainerContentChanging(ListViewBase sender, ContainerContentChangingEventArgs args)
-        {
-            if (args.InRecycleQueue)
-            {
-                args.ItemContainer.PointerEntered -= ItemContainer_PointerEntered;
-            }
-            else
-            {
-                args.ItemContainer.PointerEntered += ItemContainer_PointerEntered;
-
-                if (AnimationController.Current.IsEnableAnimation)
-                {
-                //    ProgressBar ProBar = args.ItemContainer.FindChildOfType<ProgressBar>();
-                //    Storyboard Story = new Storyboard();
-                //    DoubleAnimation Animation = new DoubleAnimation()
-                //    {
-                //        To = (args.Item as DriveDataBase).Percent,
-                //        From = 0,
-                //        EnableDependentAnimation = true,
-                //        EasingFunction = new CircleEase { EasingMode = EasingMode.EaseInOut },
-                //        Duration = new TimeSpan(0, 0, 0, 0, 800)
-                //    };
-                //    Storyboard.SetTarget(Animation, ProBar);
-                //    Storyboard.SetTargetProperty(Animation, "Value");
-                //    Story.Children.Add(Animation);
-                //    Story.Begin();
-                }
-            }
-        }
-
-        private void LibraryGrid_ContainerContentChanging(ListViewBase sender, ContainerContentChangingEventArgs args)
-        {
-            if (args.InRecycleQueue)
-            {
-                args.ItemContainer.PointerEntered -= ItemContainer_PointerEntered;
-            }
-            else
-            {
-                args.ItemContainer.PointerEntered += ItemContainer_PointerEntered;
-            }
-        }
-
-        private void ItemContainer_PointerEntered(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            if (!SettingControl.IsDoubleClickEnable)
-            {
-                if ((e.OriginalSource as FrameworkElement)?.DataContext is object Item)
-                {
-                    if (Item is LibraryFolder)
-                    {
-                        LibraryGrid.SelectedItem = Item;
-                        DeviceGrid.SelectedIndex = -1;
-                    }
-                    else if (Item is DriveDataBase)
-                    {
-                        DeviceGrid.SelectedItem = Item;
-                        LibraryGrid.SelectedIndex = -1;
-                    }
-                }
-            }
-        }
-
-        public async Task OpenTargetDriveAsync(DriveDataBase Drive)
-        {
-            switch (Drive)
-            {
-                case LockedDriveData LockedDrive:
-                    {
-                    Retry:
-                        BitlockerPasswordDialog Dialog = new BitlockerPasswordDialog();
-
-                        if (await Dialog.ShowAsync() == ContentDialogResult.Primary)
-                        {
-                            if (!await LockedDrive.UnlockAsync(Dialog.Password))
-                            {
-                                QueueContentDialog UnlockFailedDialog = new QueueContentDialog
-                                {
-                                    Title = Globalization.GetString("Common_Dialog_ErrorTitle"),
-                                    Content = Globalization.GetString("QueueDialog_UnlockBitlockerFailed_Content"),
-                                    PrimaryButtonText = Globalization.GetString("Common_Dialog_RetryButton"),
-                                    CloseButtonText = Globalization.GetString("Common_Dialog_CancelButton")
-                                };
-
-                                if (await UnlockFailedDialog.ShowAsync() == ContentDialogResult.Primary)
-                                {
-                                    goto Retry;
-                                }
-                                else
-                                {
-                                    return;
-                                }
-                            }
-
-                            StorageFolder DriveFolder = await StorageFolder.GetFolderFromPathAsync(LockedDrive.Path);
-
-                            DriveDataBase NewDrive = await DriveDataBase.CreateAsync(DriveFolder, LockedDrive.DriveType);
-
-                            if (NewDrive is LockedDriveData)
-                            {
-                                QueueContentDialog UnlockFailedDialog = new QueueContentDialog
-                                {
-                                    Title = Globalization.GetString("Common_Dialog_ErrorTitle"),
-                                    Content = Globalization.GetString("QueueDialog_UnlockBitlockerFailed_Content"),
-                                    PrimaryButtonText = Globalization.GetString("Common_Dialog_RetryButton"),
-                                    CloseButtonText = Globalization.GetString("Common_Dialog_CancelButton")
-                                };
-
-                                if (await UnlockFailedDialog.ShowAsync() == ContentDialogResult.Primary)
-                                {
-                                    goto Retry;
-                                }
-                            }
-                            else
-                            {
-                                int Index = CommonAccessCollection.DriveList.IndexOf(LockedDrive);
-
-                                if (Index >= 0)
-                                {
-                                    CommonAccessCollection.DriveList.Remove(LockedDrive);
-                                    CommonAccessCollection.DriveList.Insert(Index, NewDrive);
-                                }
-                                else
-                                {
-                                    CommonAccessCollection.DriveList.Add(NewDrive);
-                                }
-                            }
-                        }
-
-                        break;
-                    }
-                case WslDriveData:
-                case NormalDriveData:
-                    {
-                        await OpenTargetFolder(Drive.DriveFolder).ConfigureAwait(false);
-                        break;
-                    }
-            }
-        }
-
-        public async Task OpenTargetFolder(StorageFolder Folder)
-        {
-            if (Folder == null)
-            {
-                throw new ArgumentNullException(nameof(Folder), "Argument could not be null");
-            }
-
-            try
-            {
-                if (string.IsNullOrEmpty(Folder.Path))
-                {
-                    QueueContentDialog Dialog = new QueueContentDialog
-                    {
-                        Title = Globalization.GetString("Common_Dialog_TipTitle"),
-                        Content = Globalization.GetString("QueueDialog_CouldNotAccess_Content"),
-                        CloseButtonText = Globalization.GetString("Common_Dialog_CloseButton")
-                    };
-
-                    await Dialog.ShowAsync();
-                }
-                else
-                {
-                    
-                    if (CurrentPresenter != null)
-                    {
-                        await CurrentPresenter.DisplayItemsInFolder(Folder.Path);
-                    }
-                    //if (AnimationController.Current.IsEnableAnimation)
-                    //{
-                    //    Frame.Navigate(typeof(FileControl), new Tuple<WeakReference<TabViewItem>, string[]>(WeakToTabItem, new string[] { Folder.Path }), new DrillInNavigationTransitionInfo());
-                    //}
-                    //else
-                    //{
-                    //    Frame.Navigate(typeof(FileControl), new Tuple<WeakReference<TabViewItem>, string[]>(WeakToTabItem, new string[] { Folder.Path }), new SuppressNavigationTransitionInfo());
-                    //}
-                }
-            }
-            catch (Exception ex)
-            {
-                LogTracer.Log(ex, "An error was threw when entering device");
-            }
-        }
-
-        private async void DeviceGrid_DoubleTapped(object sender, Windows.UI.Xaml.Input.DoubleTappedRoutedEventArgs e)
-        {
-            LibraryGrid.SelectedIndex = -1;
-
-            if ((e.OriginalSource as FrameworkElement)?.DataContext is DriveDataBase Drive)
-            {
-                await OpenTargetDriveAsync(Drive);
-            }
-        }
-
-        private async void LibraryGrid_DoubleTapped(object sender, Windows.UI.Xaml.Input.DoubleTappedRoutedEventArgs e)
-        {
-            DeviceGrid.SelectedIndex = -1;
-
-            if ((e.OriginalSource as FrameworkElement)?.DataContext is LibraryFolder Library)
-            {
-                await OpenTargetFolder(Library.Folder).ConfigureAwait(false);
-            }
-        }
-        private void LibraryGrid_Holding(object sender, Windows.UI.Xaml.Input.HoldingRoutedEventArgs e)
-        {
-            if (e.HoldingState == HoldingState.Started)
-            {
-                if ((e.OriginalSource as FrameworkElement)?.DataContext is LibraryFolder Context)
-                {
-                    LibraryGrid.SelectedItem = Context;
-                    LibraryGrid.ContextFlyout = LibraryFlyout;
-                }
-                else
-                {
-                    LibraryGrid.ContextFlyout = null;
-                }
-            }
-        }
-
-        private void DeviceGrid_Holding(object sender, Windows.UI.Xaml.Input.HoldingRoutedEventArgs e)
-        {
-            if (e.HoldingState == HoldingState.Started)
-            {
-                if ((e.OriginalSource as FrameworkElement)?.DataContext is DriveDataBase Context)
-                {
-                    DeviceGrid.SelectedItem = Context;
-                    DeviceGrid.ContextFlyout = Context.DriveType == DriveType.Removable ? PortableDeviceFlyout : DeviceFlyout;
-                }
-                else
-                {
-                    DeviceGrid.SelectedIndex = -1;
-                    DeviceGrid.ContextFlyout = EmptyFlyout;
-                }
-            }
-        }
-
-        private async void Attribute_Click(object sender, RoutedEventArgs e)
-        {
-            DeviceInfoDialog Dialog = new DeviceInfoDialog(DeviceGrid.SelectedItem as DriveDataBase);
-            await Dialog.ShowAsync();
-        }
-
-        private void DeviceGrid_RightTapped(object sender, Windows.UI.Xaml.Input.RightTappedRoutedEventArgs e)
-        {
-            if (e.PointerDeviceType == Windows.Devices.Input.PointerDeviceType.Mouse)
-            {
-                if ((e.OriginalSource as FrameworkElement)?.DataContext is DriveDataBase Context)
-                {
-                    DeviceGrid.SelectedItem = Context;
-
-                    if (Context is LockedDriveData)
-                    {
-                        DeviceGrid.ContextFlyout = BitlockerDeviceFlyout;
-                    }
-                    else
-                    {
-                        DeviceGrid.ContextFlyout = Context.DriveType == DriveType.Removable ? PortableDeviceFlyout : DeviceFlyout;
-                    }
-                }
-                else
-                {
-                    DeviceGrid.SelectedIndex = -1;
-                    DeviceGrid.ContextFlyout = EmptyFlyout;
-                }
-            }
-        }
-
-        private async void OpenDevice_Click(object sender, RoutedEventArgs e)
-        {
-            LibraryGrid.SelectedIndex = -1;
-
-            if (DeviceGrid.SelectedItem is DriveDataBase Drive)
-            {
-                await OpenTargetDriveAsync(Drive);
-            }
-        }
-
-        private void DeviceGrid_PointerPressed(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            if ((e.OriginalSource as FrameworkElement)?.DataContext == null)
-            {
-                DeviceGrid.SelectedIndex = -1;
-            }
-            else
-            {
-                LibraryGrid.SelectedIndex = -1;
-            }
-        }
-
-        private void LibraryGrid_PointerPressed(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            if ((e.OriginalSource as FrameworkElement)?.DataContext == null)
-            {
-                LibraryGrid.SelectedIndex = -1;
-            }
-            else
-            {
-                DeviceGrid.SelectedIndex = -1;
-            }
-        }
-
-        private void Grid_PointerPressed(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            DeviceGrid.SelectedIndex = -1;
-            LibraryGrid.SelectedIndex = -1;
-        }
-
-        private void LibraryGrid_RightTapped(object sender, Windows.UI.Xaml.Input.RightTappedRoutedEventArgs e)
-        {
-            if (e.PointerDeviceType == Windows.Devices.Input.PointerDeviceType.Mouse)
-            {
-                if ((e.OriginalSource as FrameworkElement)?.DataContext is LibraryFolder Context)
-                {
-                    LibraryGrid.SelectedItem = Context;
-                    LibraryGrid.ContextFlyout = LibraryFlyout;
-                }
-                else
-                {
-                    LibraryGrid.ContextFlyout = LibraryEmptyFlyout;
-                }
-            }
-        }
-
-        private async void OpenLibrary_Click(object sender, RoutedEventArgs e)
-        {
-            DeviceGrid.SelectedIndex = -1;
-
-            if (LibraryGrid.SelectedItem is LibraryFolder Library)
-            {
-                await OpenTargetFolder(Library.Folder).ConfigureAwait(false);
-            }
-        }
-
-        private async void RemovePin_Click(object sender, RoutedEventArgs e)
-        {
-            if (LibraryGrid.SelectedItem is LibraryFolder Library)
-            {
-                CommonAccessCollection.LibraryFolderList.Remove(Library);
-                await SQLite.Current.DeleteLibraryAsync(Library.Folder.Path).ConfigureAwait(false);
-                await JumpListController.Current.RemoveItem(JumpListGroup.Library, Library.Folder).ConfigureAwait(false);
-            }
-        }
-
-        private async void LibraryProperties_Click(object sender, RoutedEventArgs e)
-        {
-            if (LibraryGrid.SelectedItem is LibraryFolder Library)
-            {
-                if (await FileSystemStorageItemBase.CreateFromStorageItemAsync(Library.Folder) is FileSystemStorageFolder Folder)
-                {
-                    await Folder.LoadMorePropertiesAsync();
-
-                    AppWindow NewWindow = await AppWindow.TryCreateAsync();
-                    NewWindow.RequestSize(new Size(420, 600));
-                    NewWindow.RequestMoveRelativeToCurrentViewContent(new Point(Window.Current.Bounds.Width / 2 - 200, Window.Current.Bounds.Height / 2 - 300));
-                    NewWindow.PersistedStateId = "Properties";
-                    NewWindow.Title = Globalization.GetString("Properties_Window_Title");
-                    NewWindow.TitleBar.ExtendsContentIntoTitleBar = true;
-                    NewWindow.TitleBar.ButtonBackgroundColor = Colors.Transparent;
-                    NewWindow.TitleBar.ButtonInactiveBackgroundColor = Colors.Transparent;
-
-                    ElementCompositionPreview.SetAppWindowContent(NewWindow, new PropertyBase(NewWindow, Folder));
-                    WindowManagementPreview.SetPreferredMinSize(NewWindow, new Size(420, 600));
-
-                    await NewWindow.TryShowAsync();
-                }
-            }
-        }
-
-        public async void Refresh_Click(object sender, RoutedEventArgs e)
-        {
-            await CommonAccessCollection.LoadDriveAsync(true);
-        }
-
-        private async void DeviceGrid_ItemClick(object sender, ItemClickEventArgs e)
-        {
-            LibraryGrid.SelectedIndex = -1;
-
-            if (!SettingControl.IsDoubleClickEnable && e.ClickedItem is DriveDataBase Drive)
-            {
-                await OpenTargetDriveAsync(Drive);
-            }
-        }
-
-        private async void LibraryGrid_ItemClick(object sender, ItemClickEventArgs e)
-        {
-            DeviceGrid.SelectedIndex = -1;
-
-            if (!SettingControl.IsDoubleClickEnable && e.ClickedItem is LibraryFolder Library)
-            {
-                await OpenTargetFolder(Library.Folder).ConfigureAwait(false);
-            }
-        }
-
-        private async void AddDevice_Click(object sender, RoutedEventArgs e)
-        {
-            FolderPicker Picker = new FolderPicker
-            {
-                SuggestedStartLocation = PickerLocationId.ComputerFolder,
-                ViewMode = PickerViewMode.Thumbnail
-            };
-            Picker.FileTypeFilter.Add("*");
-
-            StorageFolder DriveFolder = await Picker.PickSingleFolderAsync();
-
-            if (DriveFolder != null)
-            {
-                if (DriveFolder.Path.Equals(Path.GetPathRoot(DriveFolder.Path), StringComparison.OrdinalIgnoreCase) && DriveInfo.GetDrives().Where((Drive) => Drive.DriveType == DriveType.Fixed || Drive.DriveType == DriveType.Removable || Drive.DriveType == DriveType.Network).Any((Item) => Item.RootDirectory.FullName == DriveFolder.Path))
-                {
-                    if (CommonAccessCollection.DriveList.All((Item) => !Item.Path.Equals(DriveFolder.Path, StringComparison.OrdinalIgnoreCase)))
-                    {
-                        CommonAccessCollection.DriveList.Add(await DriveDataBase.CreateAsync(DriveFolder, new DriveInfo(DriveFolder.Path).DriveType));
-                    }
-                    else
-                    {
-                        QueueContentDialog Dialog = new QueueContentDialog
-                        {
-                            Title = Globalization.GetString("Common_Dialog_TipTitle"),
-                            Content = Globalization.GetString("QueueDialog_DeviceExist_Content"),
-                            CloseButtonText = Globalization.GetString("Common_Dialog_CloseButton")
-                        };
-                        await Dialog.ShowAsync();
-                    }
-                }
-                else
-                {
-                    QueueContentDialog Dialog = new QueueContentDialog
-                    {
-                        Title = Globalization.GetString("Common_Dialog_TipTitle"),
-                        Content = Globalization.GetString("QueueDialog_DeviceSelectError_Content"),
-                        CloseButtonText = Globalization.GetString("Common_Dialog_TipTitle")
-                    };
-
-                    await Dialog.ShowAsync();
-                }
-            }
-        }
-
-        public void ShowThisPC()
-        {
-            if (ThisPCView.Visibility == Visibility.Collapsed)
-            {
-                ThisPCView.Visibility = Visibility.Visible;
-                BladeViewer.Visibility = Visibility.Collapsed;
-            }
-        }
-
-        public void HideThisPC()
-        {
-            if (ThisPCView.Visibility == Visibility.Visible)
-            {
-                ThisPCView.Visibility = Visibility.Collapsed;
-                BladeViewer.Visibility = Visibility.Visible;
-            }
-        }
-
+        
     }
 }
